# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Farm Management API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "security": [
        {
            "apiKey": []
        }
    ],
    "paths": {
...
        "/farm/invitations/{invitation}": {
            "delete": {
                "operationId": "v1.farm.invitations.destroy",
                "description": "The link stops working immediately. Requires `team.invite`.\n\nErrors: `403`, `404`, `409 invitation_not_pending`.\n\n**Requires farm permission:** `team.invite`.",
                "summary": "Revoke an invitation",
                "tags": [
                    "FarmInvitation"
                ],
                "parameters": [
                    {
                        "name": "invitation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/InvitationResource"
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Invitation already accepted or revoked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string",
                                            "const": "invitation_not_pending"
                                        },
                                        "request_id": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden: account state does not allow this request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable summary."
                                        },
                                        "code": {
                                            "type": "string",
                                            "description": "Machine-readable error code.",
                                            "enum": [
                                                "account_suspended",
                                                "email_verification_required",
                                                "onboarding_required",
                                                "no_active_farm",
                                                "farm_access_denied",
                                                "forbidden"
                                            ]
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "description": "Echo of the X-Request-Id header, for support."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional structured context. Entitlement errors (feature_not_available, subscription_inactive, plan_limit_reached) carry entitlement_key and, for limits, limit, usage and remaining."
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found (or it belongs to another farm)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable summary."
                                        },
                                        "code": {
                                            "type": "string",
                                            "description": "Machine-readable error code.",
                                            "enum": [
                                                "not_found"
                                            ]
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "description": "Echo of the X-Request-Id header, for support."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional structured context. Entitlement errors (feature_not_available, subscription_inactive, plan_limit_reached) carry entitlement_key and, for limits, limit, usage and remaining."
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    },
                    "419": {
                        "description": "CSRF token missing or expired: call GET /auth/csrf-cookie and retry",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable summary."
                                        },
                                        "code": {
                                            "type": "string",
                                            "description": "Machine-readable error code.",
                                            "enum": [
                                                "session_expired"
                                            ]
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "description": "Echo of the X-Request-Id header, for support."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional structured context. Entitlement errors (feature_not_available, subscription_inactive, plan_limit_reached) carry entitlement_key and, for limits, limit, usage and remaining."
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/invitations/accept": {
            "post": {
                "operationId": "v1.invitations.accept",
                "description": "Body: `{ \"token\": \"<token from the emailed link>\" }`. The caller must be signed in with a VERIFIED\naccount whose email equals the invited email. Creates (or re-activates) the membership with the invited\nrole; a user who has not set up a farm yet skips farm setup (`next_action = \"none\"`). The response is the\nauth state; its `farm` is your default farm - send `X-Farm-Id` with `meta.accepted_farm_id` to work on the newly joined farm if you already belong to another one.\n\nErrors: `401`, `403 email_verification_required` / `invitation_email_mismatch`, `404 invitation_not_found`,\n`409 already_member`, `410 invitation_used` / `invitation_revoked` / `invitation_expired`, `422`, `429`.",
                "summary": "Accept a farm invitation",
                "tags": [
                    "InvitationAcceptance"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AcceptInvitationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "meta": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Invitation accepted."
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "410": {
                        "description": "Invitation used, revoked or expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string",
                                            "enum": [
                                                "invitation_used",
                                                "invitation_revoked",
                                                "invitation_expired"
                                            ]
                                        },
                                        "request_id": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Already a member",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string",
                                            "const": "already_member"
                                        },
                                        "request_id": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown invitation token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string",
                                            "const": "invitation_not_found"
                                        },
                                        "request_id": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden: account state does not allow this request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable summary."
                                        },
                                        "code": {
                                            "type": "string",
                                            "description": "Machine-readable error code.",
                                            "enum": [
                                                "account_suspended",
                                                "email_verification_required"
                                            ]
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "description": "Echo of the X-Request-Id header, for support."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional structured context. Entitlement errors (feature_not_available, subscription_inactive, plan_limit_reached) carry entitlement_key and, for limits, limit, usage and remaining."
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limited (see Retry-After header)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable summary."
                                        },
                                        "code": {
                                            "type": "string",
                                            "description": "Machine-readable error code.",
                                            "enum": [
                                                "too_many_requests"
                                            ]
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "description": "Echo of the X-Request-Id header, for support."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional structured context. Entitlement errors (feature_not_available, subscription_inactive, plan_limit_reached) carry entitlement_key and, for limits, limit, usage and remaining."
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    },
                    "419": {
                        "description": "CSRF token missing or expired: call GET /auth/csrf-cookie and retry",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable summary."
                                        },
                                        "code": {
                                            "type": "string",
                                            "description": "Machine-readable error code.",
                                            "enum": [
                                                "session_expired"
                                            ]
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "description": "Echo of the X-Request-Id header, for support."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional structured context. Entitlement errors (feature_not_available, subscription_inactive, plan_limit_reached) carry entitlement_key and, for limits, limit, usage and remaining."
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Sanctum session cookie set by /auth/login, /auth/register or /auth/google. Send requests with credentials and, for POST/PUT/PATCH/DELETE, the X-XSRF-TOKEN header (see docs/api/README.md).",
                "in": "cookie",
                "name": "farm_management_api_session"
            }
        },
        "schemas": {
            "AcceptInvitationRequest": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "maxLength": 128
                    }
                },
                "required": [
                    "token"
                ],
                "title": "AcceptInvitationRequest"
            },
...
            "InvitationResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "role_label": {
                        "type": "string",
                        "enum": [
                            "Owner",
                            "Manager",
                            "Farm Worker",
                            "Finance"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "accepted",
                            "revoked",
                            "expired",
                            "pending"
                        ]
                    },
                    "expires_at": {
                        "type": "string"
                    },
                    "last_sent_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "invited_by": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "email": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "id",
                            "name",
                            "email"
                        ]
                    }
                },
                "required": [
                    "id",
                    "email",
                    "role",
                    "role_label",
                    "status",
                    "expires_at",
                    "last_sent_at",
                    "created_at",
                    "invited_by"
                ],
                "title": "InvitationResource"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}